Make a StoreReceiptService
